Allow multiple, stackable modals
Modals are now stacked above each other if you call them from
another modal and the user has JavaScript activated. This allows
to guide the user through a process of several modals in a better
way.

There still need to be more tests and exampels written for that
specific case. Also browser tests need to be done.
The concept on mobile isn't yet fully developed and should be
redefined.

Closes #20.