Fix close button position
The close button didn't stick to it's position after hitting
the first media-query towards mobile. The mobile close button
is not affected by this.
Caused by adjusting styles within this certain mq directly on
the .modal-close element rather on it's pseudo :after element
the correct values have now been applied.
The close button should now stick on it's position when the
module resizes.