Use jQuery as CustomEvent fallback instead of Bean
This commit removes Bean from CSS Modal's IE8 dependency list and
introduces jQuery as this dependency for CustomEvents.
CustomEventus can be mocked in IE8 with some difficulties - this should
not be the problem of us as implementors though which is why we use a
library for correct behavior. In the past we've used bean as a
lightwight solution.

Nearly all webprojects these days require jQuery for their JavaScript
which is why I think jQuery is more robust and easier to use for most
developers.
You get full (JS) IE8 compatibility by simply including jQuery into your
project.