Gallery: Add possibility to use iframes in slides
A gallery can now handle iframes: Just provide a link to the according
iframe in `data-src-fullsize` instead of an image and set `data-iframe`
on the enclosing link.

    <a href="#modal-gallery/iframe" data-iframe>
        <img alt="" src="image.jpg" data-src-fullsize="//iframe-url.com/">
    </a>

This adds functionality for videos from Vimeo and Youtube. They are not
paused when skipping to the next/prev element. Please consider adding a
custom script for handling this behavior.

The defined max-width of an iframe is currently 600px. You can change
this by setting a variable `$gallery--iframe-size` which overwrites the
current default value.