Fix bug on IE11 with close button and overlay of modal
In IE11 the modal overlay some content because of the element defined as
modal:before in the mobile view. Thus one could not click elements that
underlay this part of the page.

The commit also delegates some more styling of the close element into
close itself instead of the modal's pseudo elements.